Adding UI and Validation

+++ src/main/java/springBootExample/controller/UserController.java
@@ -1,18 +1,20 @@
 package springBootExample.controller;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.web.bind.annotation.*;
 import springBootExample.model.entity.User;
 import springBootExample.model.json.*;
+import springBootExample.model.repository.UserMapper;
 import springBootExample.model.repository.UserRepository;
 
 import javax.validation.ConstraintViolation;
 import javax.validation.Validation;
 import javax.validation.Validator;
 import javax.validation.ValidatorFactory;
-import java.util.Iterator;
+import java.util.ArrayList;
 import java.util.Set;
 
 /**
@@ -31,27 +33,41 @@ public class UserController {
 
     private Validator validator = null;
 
+    private final Logger log = LoggerFactory.getLogger(this.getClass());
+
     public UserController() {
         ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
         validator = factory.getValidator();
     }
 
     /**
-
      * @param id
      * @return JsonResponse
      */
     @RequestMapping(method = RequestMethod.GET, value = "{id}")
     JsonResponse getUser(@PathVariable String id) {
         ResponseManager rm = new ResponseManager();
-        User entity = userRepository.findOne(Long.parseLong(id));
 
-        if (entity == null) {
-            rm.addError("User does not exist.");
+        if (!userRepository.userExists(id)) {
+            rm.addError("general : User does not exist.");
             return rm.getResponse();
         }
 
-        return userRepository.getResponseFromUser(entity);
+        return UserMapper.INSTANCE.userToUserDto(userRepository.findOne(Long.parseLong(id)));
+    }
+
+    /**
+     * @return Iterable
+     */
+    @RequestMapping(method = RequestMethod.GET)
+    Iterable<UserDto> getUsers() {
+        ArrayList<UserDto> result = new ArrayList<>();
+
+        userRepository.findAll().forEach(
+                entity -> result.add(UserMapper.INSTANCE.userToUserDto(entity))
+        );
+
+        return result;
     }
 
     /**
@@ -63,20 +79,25 @@ public class UserController {
     @ResponseBody
     JsonResponse addUser(@RequestBody UserDto user) {
         ResponseManager rm = new ResponseManager();
-        Set<ConstraintViolation<UserDto>> constraintViolations = validator.validate(user);
 
-        if( constraintViolations.size() > 0 ) {
-            Iterator<ConstraintViolation<UserDto>> iterator = constraintViolations.iterator();
-            while(iterator.hasNext()) {
-                ConstraintViolation<UserDto> error = iterator.next();
-                rm.addError(error.getPropertyPath().toString() +  " : " + error.getMessage());
+        Set<ConstraintViolation<UserDto>> constraintViolations = validator.validate(user);
+        if (constraintViolations.size() > 0) {
+            for (ConstraintViolation<UserDto> error : constraintViolations) {
+                rm.addError(error.getPropertyPath().toString() + " : " + error.getMessage());
             }
             return rm.getResponse();
         }
 
-        User entity = userRepository.createEntityFromResponse(user);
-        userRepository.save(entity);
-        return userRepository.getResponseFromUser(entity);
+        User entity = UserMapper.INSTANCE.userDtoToUser(user);
+
+        try {
+            userRepository.save(entity);
+            return UserMapper.INSTANCE.userToUserDto(entity);
+        } catch (Exception e) {
+            rm.addError("general : Something went wrong, please contact admin.");
+            log.error(e.toString());
+        }
+        return rm.getResponse();
     }
 
     /**
@@ -89,16 +110,30 @@ public class UserController {
     @ResponseBody
     JsonResponse update(@PathVariable String id, @RequestBody UserDto user) {
         ResponseManager rm = new ResponseManager();
-        User entity = userRepository.findOne(Long.parseLong(id));
 
-        if (entity == null) {
-            rm.addError("User does not exist.");
+        if (!userRepository.userExists(id)) {
+            rm.addError("general : User does not exist.");
             return rm.getResponse();
         }
-        //TODO add validation.
-        userRepository.updateEntityFromResponse(user, entity);
-        userRepository.save(entity);
-        return userRepository.getResponseFromUser(entity);
+
+        Set<ConstraintViolation<UserDto>> constraintViolations = validator.validate(user);
+        if (constraintViolations.size() > 0) {
+            for (ConstraintViolation<UserDto> error : constraintViolations) {
+                rm.addError(error.getPropertyPath().toString() + " : " + error.getMessage());
+            }
+            return rm.getResponse();
+        }
+
+        User entity = UserMapper.INSTANCE.userDtoToUser(user);
+
+        try {
+            userRepository.save(entity);
+            return UserMapper.INSTANCE.userToUserDto(entity);
+        } catch (Exception e) {
+            rm.addError("general : Something went wrong, please contact admin.");
+            log.error(e.toString());
+        }
+        return rm.getResponse();
     }
 
     /**
@@ -111,16 +146,17 @@ public class UserController {
     JsonResponse deleteUser(@PathVariable String id) {
         ResponseManager rm = new ResponseManager();
 
-        //TODO replace try catch with validation
+        if (!userRepository.userExists(id)) {
+            rm.addError("general : User does not exist.");
+            return rm.getResponse();
+        }
+
         try {
             userRepository.delete(Long.parseLong(id));
-        } catch (EmptyResultDataAccessException e) {
-            rm.addError("User does not exist.");
         } catch (Exception e) {
-            rm.addError("Something went wrong, please contact admin.");
+            rm.addError("general : Something went wrong, please contact admin.");
+            log.error(e.toString());
         }
         return rm.getResponse();
     }
-    //TODO add exception logging.
-    //TODO figure out error messaging pattern or s.th
 }
